src/requirements.txt
python-dotenv
pandas
pymongo
sqlalchemy
